+++ satellite/console/consoleauth/apikey.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package auth
+package consoleauth
 
 import (
 	"context"
