+++ satellite/console/consoleweb/server.go
@@ -31,9 +31,9 @@ import (
 
 	"storj.io/common/errs2"
 	"storj.io/common/uuid"
-	"storj.io/storj/pkg/auth"
 	"storj.io/storj/private/web"
 	"storj.io/storj/satellite/console"
+	"storj.io/storj/satellite/console/consoleauth"
 	"storj.io/storj/satellite/console/consoleweb/consoleapi"
 	"storj.io/storj/satellite/console/consoleweb/consoleql"
 	"storj.io/storj/satellite/console/consoleweb/consolewebauth"
@@ -319,7 +319,7 @@ func (server *Server) withAuth(handler http.Handler) http.Handler {
 				return console.WithAuthFailure(ctx, err)
 			}
 
-			ctx = auth.WithAPIKey(ctx, []byte(token))
+			ctx = consoleauth.WithAPIKey(ctx, []byte(token))
 
 			auth, err := server.service.Authorize(ctx)
 			if err != nil {
@@ -354,7 +354,7 @@ func (server *Server) bucketUsageReportHandler(w http.ResponseWriter, r *http.Re
 		return
 	}
 
-	auth, err := server.service.Authorize(auth.WithAPIKey(ctx, []byte(token)))
+	auth, err := server.service.Authorize(consoleauth.WithAPIKey(ctx, []byte(token)))
 	if err != nil {
 		server.serveError(w, http.StatusUnauthorized)
 		return
